isys & psys package lib2600b0 for scci_master_20161129_1224
Releasing on scci_master_20161123_1219

906d927 [HSD1405485256] Firmware error exposed by moving to
win10 WDK 2016-11-28
7d96969 [IPU5][SIS] Work-around due to HSD1804337130 2016-11-28
6c05568 [IPU_HOST] Updated sync function & adds doxygen
2016-11-23
fb691b0 [IPU_HOST] extend subsystem sync to multiple frame
2016-11-28
3f9ce25 [HSD1804670855] Removal of realpath to avoid IDK
compilation error 2016-11-28
7308c38 [IPU5][PREGDC] Add release notes to release package
2016-11-28
bf3fd0d [IPU5] Update config files with flags for external
memories 2016-11-25
739583b [PnP] Log collection into TC artifacts 2016-11-27
99510b1 [IPU5] Add LB VMEM to device properties 2016-11-25
731d6b4 [HSD1304066121] guard fixing isp kernels c-files (part
4) 2016-11-25
bb25964 [HSD1304066121] guard fixing isp c-files part 1
2016-11-25
df17e8b [PG gaussian_mem] Clean haaf file step 1 2016-11-25
5594bde [IPU5_ISL-PS]Updating the ISL_PS release notes for GLV.
2016-11-25
dcf654a [IDK] Remove IDK specific run script from FW trunk
2016-11-25
9f664bf [EQ] Return value from tests is corrected 2016-11-25
da7e692 [IPU5] increase envelope for 8 bit input only gdc
2016-11-25
1f7d4d3 [MBR NCI] Add MBR NCI SP test 2016-11-08
9a7dc92 [BUILD] migrating ipu5 PG tests to ipu_host 2016-09-30
fb2d3ba [VPOST|IPU5] Use 8b input for vpostgdc 2016-11-16
5ca576e [FWAPI] Remove isys_host from host_app 2016-11-17
01eba12 [TOOLS #JIFAA-895] Simulator warning finder 2016-11-24
1fd91aa [FWAPI] Isys multi-session 2016-11-24
56f96b3 [IDK] Create hostapp run_test script 2016-11-21
f5b18d4 [IDK] Introduce smoke_lite field in examples.json
2016-11-24
d301cb5 [IoTG] Vertical cropping support for IoTG kernels
2016-09-10
7893c85 [HSD1304066121] guard fixing isp c-files in bxt_sandbox
2016-11-24
d024d21 [GENPIPE] Fix precondition issue in Genpipe 2016-11-24
952b7ad [IPU5] Remove DAI DMEM usages 2016-11-24
4e730e5 [IPU5] Cleanup Kernel info 2016-11-24
b045a38 [TNR5_3] Add process iterator for tnr5_3 2016-11-22
3a8d70c [TNR5_3] Common search area check func for MVD 2016-11-22
d75fa2f [TNR5_3] Add common code/functions for MVD/FMVE
2016-11-22
fa461b1 [bufqueue] Add an option to hoist out acquire and
dequeue 2016-09-26
417cc3c [FWQ] Update parametrs for vhdr_pp_1920x1088 8x tests
2016-11-17
1c63c68 [CNLB0 MBR] Adding test for Lanczos filter 2016-11-22
3670443 [HSD1405431804] Updating ref bin path in haaf 2016-11-23
53ec515 [cnl vpregdc_hq_vcr2_dpcm] added 1920x1088 sched only
test temp 2016-11-22
a005a56 [FWQ] Fix the bxt_height for a paf test 2016-11-23
ac1662b [FWQ] Port CNL PAF tests to use pixter 2016-11-21
eba15d9 [HSD1304066121] guard fixing for tproxy 2016-11-23
c527617 [IPU5] Remove MBR test from smoke_fpga 2016-11-23
271385e [IPU5][PREGDC] Update release notes of PREGDC program
groups 2016-11-23
af5ef92 [GENPIPE][GENREADER] Check that encode functions return
success 2016-11-18
b561df0 [FWAPI] Fix Psys bugs related to pkg_dir 2016-11-18
40ed6a9 [HSD1304066121] guard fixing ktb tests (part 6)
2016-11-23
39761a5 [cnlB0] Enable profile collection for CNL postgdc PG's
2016-11-22
a985af1 [cnl smoke smoe_fpga] add performance tests in
smoke/smoke_fpga 2016-11-21
d0163b7 [GLV-ISL-PS] Removing symlinks for dev test binaries
2016-11-22
c91938e [TNR5_21] Post GDC KTB test for 4K with NS bypass
2016-11-16
f2f0bbb [Multi_threading] extend multi_thread example - frame
independent 2016-11-22
0607c90 [TNR5_3] Shfl idx to extract 8x8 center from 16x16
2016-11-21
0922884 [on_the_fly] ISYS changes and test for on_the_fly
2016-11-03
84e799e [Devproxy][ISA] Remove an uninitialized read 2016-11-22
1a8dc27 [STNR5.2.1] Preparation for single breakout - recsim
2016-11-21
cdb7676 [STNR5.21] Recsim Put 2016-11-16
f243c3d [bxt_sandbox|JIPAI-497] Fixed hardcoded /p/siliconhive
paths 2016-11-16
fccfde0 [test_executor] Return 1 if no tests are found 2016-11-17
f53f8ea [ISYS] FW-based WA for HSD 1304553438 2016-11-04

FW OTM: tag scci_master_20161129_1224

+++ drivers/media/pci/intel-ipu4/ipu4_bxtB0/lib2600psys/lib/psysapi/param/src/ia_css_program_group_param.c
@@ -515,7 +515,7 @@ ia_css_terminal_param_init(ia_css_terminal_param_t *terminal_param,
 
 	terminal_param->size = sizeof(ia_css_terminal_param_t);
 	/* parent is at negative offset from current program.*/
-	terminal_param->parent_offset = -offset;
+	terminal_param->parent_offset = -((int32_t)offset);
 	/*TODO: Kernel_bitmap setting. ?*/
 	terminal_param->frame_format_type = frame_format_type;
